Move structure helpers to structure module
