+++ src/pyscaffold/api/helpers.py
@@ -2,248 +2,15 @@
 Useful functions for manipulating the action list and project structure.
 """
 
-from copy import deepcopy
-from pathlib import PurePath
-
 from ..exceptions import ActionNotFound
 from ..identification import get_id
 from ..log import logger
-from ..operations import create
-from ..structure import define_structure, structure_leaf
+from ..structure import define_structure
 
 logger = logger  # Sphinx workaround to force documenting imported members
 """Logger wrapper, that provides methods like :obj:`~.ReportLogger.report`.
 See :class:`~.ReportLogger`.
 """
-
-
-# -------- Project Structure --------
-
-
-def modify(struct, path, modifier):
-    """Modify the contents of a file in the representation of the project tree.
-
-    If the given path, does not exist the parent directories are automatically
-    created.
-
-    Args:
-        struct (dict): project representation as (possibly) nested
-            :obj:`dict`. See :obj:`~.merge`.
-
-        path (os.PathLike): path-like string or object relative to the
-            structure root. The following examples are equivalent::
-
-                from pathlib import Path
-
-                'docs/api/index.html'
-                Path('docs', 'api', 'index.html')
-
-            .. versionchanged:: 4.0
-               The function no longer accepts a list of strings of path parts.
-
-        modifier (callable): function (or callable object) that receives the
-            old content and the old file operation as arguments and returns
-            a tuple with the new content and new file operation.
-            Note that, if the file does not exist in ``struct``, ``None`` will
-            be passed as argument. Example::
-
-                modifier = lambda old, op: ((old or '') + 'APPENDED CONTENT'!, op)
-                modifier = lambda old, op: ('PREPENDED CONTENT!' + (old or ''), op)
-
-            .. versionchanged:: 4.0
-               ``modifier`` requires 2 arguments and now is a mandatory argument.
-
-    .. versionchanged:: 4.0
-       ``update_rule`` is no longer an argument. Instead the arity ``modifier`` was
-       changed to accept 2 arguments instead of only 1. This is more suitable to
-       handling the new :obj:`pyscaffold.operations` API.
-
-    Returns:
-        dict: updated project tree representation
-
-    Note:
-        Use an empty string as content to ensure a file is created empty
-        (``None`` contents will not be created).
-    """
-    # Retrieve a list of parts from a path-like object
-    path_parts = PurePath(path).parts
-
-    # Walk the entire path, creating parents if necessary.
-    root = deepcopy(struct)
-    last_parent = root
-    name = path_parts[-1]
-    for parent in path_parts[:-1]:
-        last_parent = last_parent.setdefault(parent, {})
-
-    # Get the old value if existent.
-    old_value = structure_leaf(last_parent.get(name))
-
-    # Update the value.
-    new_value = modifier(*old_value)
-    last_parent[name] = _merge_file_leaf(old_value, new_value)
-
-    return root
-
-
-def ensure(struct, path, content=None, file_op=create):
-    """Ensure a file exists in the representation of the project tree
-    with the provided content.
-    All the parent directories are automatically created.
-
-    Args:
-        struct (dict): project representation as (possibly) nested
-            :obj:`dict`. See :obj:`~.merge`.
-
-        path (os.PathLike): path-like string or object relative to the
-            structure root. See :obj:`~.modify`.
-
-            .. versionchanged:: 4.0
-               The function no longer accepts a list of strings of path parts.
-
-        content (str): file text contents, ``None`` by default.
-            The old content is preserved if ``None``.
-
-        file_op: see :obj:`pyscaffold.operations`, :obj:`~.create`` by default.
-
-    .. versionchanged:: 4.0
-       Instead of a ``update_rule`` flag, the function now accepts a :obj:`file_op
-       <pyscaffold.oprtations.FileOp>`.
-
-    Returns:
-        dict: updated project tree representation
-
-    Note:
-        Use an empty string as content to ensure a file is created empty.
-    """
-    return modify(
-        struct, path, lambda old, _: (old if content is None else content, file_op)
-    )
-
-
-def reject(struct, path):
-    """Remove a file from the project tree representation if existent.
-
-    Args:
-        struct (dict): project representation as (possibly) nested
-            :obj:`dict`. See :obj:`~.merge`.
-
-        path (os.PathLike): path-like string or object relative to the
-            structure root. See :obj:`~.modify`.
-
-            .. versionchanged:: 4.0
-               The function no longer accepts a list of strings of path parts.
-
-    Returns:
-        dict: modified project tree representation
-    """
-    # Retrieve a list of parts from a path-like object
-    path_parts = PurePath(path).parts
-
-    # Walk the entire path, creating parents if necessary.
-    root = deepcopy(struct)
-    last_parent = root
-    name = path_parts[-1]
-    for parent in path_parts[:-1]:
-        if parent not in last_parent:
-            return root  # one ancestor already does not exist, do nothing
-        last_parent = last_parent[parent]
-
-    if name in last_parent:
-        del last_parent[name]
-
-    return root
-
-
-def merge(old, new):
-    """Merge two dict representations for the directory structure.
-
-    Basically a deep dictionary merge, except from the leaf update method.
-
-    Args:
-        old (dict): directory descriptor that takes low precedence
-                    during the merge
-        new (dict): directory descriptor that takes high precedence
-                    during the merge
-
-    The directory tree is represented as a (possibly nested) dictionary.
-    The keys indicate the path where a file will be generated, while the
-    value indicates the content.  Additionally, tuple values are allowed in
-    order to specify the rule that will be followed during an ``update``
-    operation (see :obj:`pyscaffold.operations`). In this case, the first element
-    is the file content and the second element is the update rule.
-    For example, the dictionary::
-
-        {'namespace': {
-            'module.py': ('print("Hello World!")', no_overwrite())}}
-
-    represents a ``namespace/module.py`` file inside the project folder
-    with content ``print("Hello World!")``, that will be created only if not
-    present.
-
-    .. versionchanged:: 4.0
-        Project structure now considers everything **under** the
-        top level project folder.
-
-    Returns:
-        dict: resulting merged directory representation
-
-    Note:
-        Use an empty string as content to ensure a file is created empty.
-        (``None`` contents will not be created).
-    """
-    return _inplace_merge(deepcopy(old), new)
-
-
-def _inplace_merge(old, new):
-    """Similar to :obj:`~.merge` but modifies the first dict."""
-
-    for key, value in new.items():
-        old_value = old.get(key, None)
-        new_is_dict = isinstance(value, dict)
-        old_is_dict = isinstance(old_value, dict)
-        if new_is_dict and old_is_dict:
-            old[key] = _inplace_merge(old_value, value)
-        elif old_value is not None and not new_is_dict and not old_is_dict:
-            # both are defined and final leaves
-            old[key] = _merge_file_leaf(old_value, value)
-        else:
-            old[key] = deepcopy(value)
-
-    return old
-
-
-def _merge_file_leaf(old_value, new_value):
-    """Merge leaf values for the directory tree representation.
-
-    The leaf value is expected to be a tuple ``(content, update_rule)``.
-    When a string is passed, it is assumed to be the content and
-    ``None`` is used for the update rule.
-
-    Args:
-        old_value (tuple or str): descriptor for the file that takes low
-                                  precedence during the merge
-        new_value (tuple or str): descriptor for the file that takes high
-                                  precedence during the merge
-
-    Note:
-        ``None`` contents are ignored, use and empty string to force empty
-        contents.
-
-    Returns:
-        tuple or str: resulting value for the merged leaf
-    """
-    if not isinstance(old_value, (list, tuple)):
-        old_value = (old_value, None)
-    if not isinstance(new_value, (list, tuple)):
-        new_value = (new_value, None)
-
-    content = new_value[0] if new_value[0] is not None else old_value[0]
-    rule = new_value[1] if new_value[1] is not None else old_value[1]
-
-    if rule is None:
-        return content
-
-    return (content, rule)
 
 
 # -------- Action List --------
