+++ src/pyscaffold/extensions/tox.py
@@ -2,7 +2,7 @@
 Extension that generates configuration files for the Tox test automation tool.
 """
 
-from ..api import helpers
+from .. import structure
 from ..operations import no_overwrite
 from ..templates import get_template
 from . import Extension
@@ -36,4 +36,4 @@ class Tox(Extension):
         """
         files = {"tox.ini": (get_template("tox_ini"), no_overwrite())}
 
-        return helpers.merge(struct, files), opts
+        return structure.merge(struct, files), opts
