+++ tests/api/test_helpers.py
@@ -1,159 +1,9 @@
-#!/usr/bin/env python
-from pathlib import PurePath as Path
-
 import pytest
 
-from pyscaffold import actions, operations
+from pyscaffold import actions
 from pyscaffold.api import helpers
 from pyscaffold.exceptions import ActionNotFound
 from pyscaffold.structure import define_structure
-
-NO_OVERWRITE = operations.no_overwrite()
-SKIP_ON_UPDATE = operations.skip_on_update()
-
-
-def test_merge_basics():
-    # Given an existing structure,
-    structure = {"a": {"b": {"c": "1", "d": "2"}}}
-    # when it is merged to another structure with some common folder
-    extra_files = {"a": {"b": {"c": "0"}, "e": "2"}, "f": {"g": {"h": "0"}}}
-    structure = helpers.merge(structure, extra_files)
-    # then the result, should contain both files from the original and the
-    # merged structure,
-    assert structure["a"]["b"]["d"] == "2"
-    assert structure["f"]["g"]["h"] == "0"
-    assert structure["a"]["e"] == "2"
-    # the common leaves should be overridden and a tuple (content, rule)
-    assert structure["a"]["b"]["c"] == "0"
-
-
-def test_merge_rules_just_in_original():
-    # When an update rule exists in the original structure,
-    structure = {"a": {"b": ("0", SKIP_ON_UPDATE)}}
-    # but not in the merged,
-    extra_files = {"a": {"b": "3"}}
-    structure = helpers.merge(structure, extra_files)
-    # then just the content should be updated
-    # and the rule should be kept identical
-    assert structure["a"]["b"] == ("3", SKIP_ON_UPDATE)
-
-
-def test_merge_rules_just_in_merged():
-    # When an update rule does not exist in the original structure,
-    structure = {"a": {"b": "0"}}
-    # but exists in the merged,
-    extra_files = {"a": {"b": (None, SKIP_ON_UPDATE)}}
-    structure = helpers.merge(structure, extra_files)
-    # then just the rule should be updated
-    # and the content should be kept identical
-    assert structure["a"]["b"] == ("0", SKIP_ON_UPDATE)
-
-
-def test_empty_string_leads_to_empty_file_during_merge():
-    # When the original structure contains a leaf,
-    structure = {"a": {"b": "0"}}
-    # and the merged structure overrides it with an empty content
-    extra_files = {"a": {"b": ""}}
-    structure = helpers.merge(structure, extra_files)
-    # then the resulting content should exist and be empty
-    assert structure["a"]["b"] == ""
-
-
-def test_modify_non_existent():
-    # Given the original structure does not contain a leaf
-    # that is targeted by the modify method,
-    structure = {"a": {"b": "0"}}
-
-    # When the modify is called
-    # Then the argument passed should be None
-    def _modifier(old, op):
-        assert old is None
-        return ("1", op)
-
-    structure = helpers.modify(structure, Path("a", "c"), _modifier)
-
-    # But the result of the modifier function should be included in the tree
-    assert structure["a"]["c"] == ("1", operations.create)
-
-
-def test_modify_file_op():
-    structure = {"a": {"b": "0"}}
-    structure = helpers.modify(structure, "a/b", lambda text, _: (text, SKIP_ON_UPDATE))
-    assert structure["a"]["b"] == ("0", SKIP_ON_UPDATE)
-
-
-def test_ensure_nested():
-    # When the original structure does not contain a leaf
-    structure = {"a": {"b": "0"}}
-    # that is added using the ensure method,
-    structure = helpers.ensure(structure, Path("a", "c", "d", "e", "f"), content="1")
-    # then all the necessary parent folder should be included
-    assert isinstance(structure["a"]["c"], dict)
-    assert isinstance(structure["a"]["c"]["d"], dict)
-    assert isinstance(structure["a"]["c"]["d"]["e"], dict)
-    # and the file itself
-    assert structure["a"]["c"]["d"]["e"]["f"] == ("1", operations.create)
-
-
-def test_ensure_overriden():
-    # When the original structure contains a leaf
-    structure = {"a": {"b": "0"}}
-    # that is overridden using the ensure method,
-    structure = helpers.ensure(structure, Path("a", "b"), content="1")
-    # and the file content should be overridden
-    assert structure["a"]["b"] == ("1", operations.create)
-
-
-def test_ensure_path():
-    # When the ensure method is called with an string path
-    structure = {}
-    structure = helpers.ensure(structure, "a/b/c/d", content="1")
-    # Then the effect should be the same as if it were split
-    assert structure["a"]["b"]["c"]["d"] == ("1", operations.create)
-
-
-def test_ensure_file_op():
-    # When the ensure method is called with a file_op and no content
-    structure = {"a": {"b": "0"}}
-    structure = helpers.ensure(structure, "a/b", file_op=NO_OVERWRITE)
-    # Then the content should remain the same
-    # But the file_op should change
-    assert structure["a"]["b"] == ("0", NO_OVERWRITE)
-
-
-def test_reject():
-    # When the original structure contain a leaf
-    structure = {"a": {"b": {"c": "0"}}}
-    # that is removed using the reject method,
-    structure = helpers.reject(structure, Path("a", "b", "c"))
-    # then the structure should not contain the file
-    assert "c" not in structure["a"]["b"]
-
-
-def test_reject_without_ancestor():
-    # Given a defined structure,
-    structure = {"a": {"b": {"c": "0"}}}
-    # when someone tries to remove a file using the reject method
-    # but one of its ancestor does not exist in the structure,
-    structure = helpers.reject(structure, "a/b/x/c")
-    # then the structure should be the same
-    assert structure["a"]["b"]["c"] == "0"
-    assert len(structure["a"]["b"]["c"]) == 1
-    assert len(structure["a"]["b"]) == 1
-    assert len(structure["a"]) == 1
-
-
-def test_reject_without_file():
-    # Given a defined structure,
-    structure = {"a": {"b": {"c": "0"}}}
-    # when someone tries to remove a file using the reject method
-    # but one of its ancestor does not exist in the structure,
-    structure = helpers.reject(structure, "a/b/x")
-    # then the structure should be the same
-    assert structure["a"]["b"]["c"] == "0"
-    assert len(structure["a"]["b"]["c"]) == 1
-    assert len(structure["a"]["b"]) == 1
-    assert len(structure["a"]) == 1
 
 
 def custom_action(structure, options):
